update models path
