+++ alembic/versions/d78a1b3c7737_.py
@@ -1,0 +1,30 @@
+"""empty message
+
+Revision ID: d78a1b3c7737
+Revises: ffe05e8421a7
+Create Date: 2024-01-26 17:48:35.706408
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'd78a1b3c7737'
+down_revision: Union[str, None] = 'ffe05e8421a7'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('users', sa.Column('bio', sa.String(length=255), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('users', 'bio')
+    # ### end Alembic commands ###
