+++ app/models/__init__.py
@@ -1,0 +1,4 @@
+from app.db.base import Base
+
+from app.models.user import *
+from app.models.blog import *
