+++ app/models/user.py
@@ -1,0 +1,22 @@
+import uuid
+from sqlalchemy import Column, String, String, Column, DateTime, func
+from sqlalchemy.orm import relationship
+from sqlalchemy.dialects.postgresql import UUID
+
+from app.db.base import Base
+
+
+class User(Base):
+    __tablename__ = 'users'
+    __table_args__ = {'extend_existing': True}
+
+    id = Column(UUID(as_uuid=True), default=uuid.uuid4, primary_key=True, index=True)
+    first_name = Column(String(50), index=True, nullable=False)
+    last_name = Column(String(50), index=True, nullable=False)
+    email = Column(String(50), unique=True, index=True, nullable=False)
+    password = Column(String(255), nullable=False)
+    bio = Column(String(255))
+    created_at = Column(DateTime, default=func.now())
+    updated_at = Column(DateTime, default=func.now(), onupdate=func.now())
+
+    blogs = relationship('Blog', back_populates='user')
