+++ main.py
@@ -7,11 +7,11 @@ from fastapi.params import Depends
 from sqlalchemy.orm import Session
 
 from logger import logger
-from models import User, Blog
+from app.models import User, Blog
 from app.schemas import *
 from app.db.base import engine, get_db
 
-from models import Base
+from app.models import Base
 
 
 @asynccontextmanager
